DRF/museum/requirements.txt
django
djangorestframework
requests
pillow
celery
redis